Add impls of AsRef<str> for Elevation and From<Elevation> for Classes
Moved Card example to doc.rs
