+++ src/card/mod.rs
@@ -1,6 +1,9 @@
+#[cfg(feature = "doc")]
+pub mod doc;
+
+use crate::elevation::Elevation;
 use yew::prelude::*;
 use yew::Classes;
-use crate::elevation::Elevation;
 
 #[derive(Clone, PartialEq, Properties)]
 pub struct CardProps {
@@ -41,52 +44,17 @@ impl Component for Card {
     }
 
     fn view(&self) -> Html {
-        let mut class = Classes::from("bp3-card")
-            .extend(&self.props.class);
-        class.push(self.props.elevation.as_css_class());
+        let mut classes = Classes::from("bp3-card")
+            .extend(&self.props.class)
+            .extend(self.elevation);
         if self.props.interactive {
-            class.push("bp3-interactive");
+            classes.push("bp3-interactive");
         }
 
         html! {
             <div class=class onclick={self.props.onclick.clone()}>
                 {self.props.children.clone()}
             </div>
-        }
-    }
-}
-
-#[cfg(feature = "doc")]
-pub mod doc {
-    use super::*;
-
-    pub struct CardDoc {}
-
-    impl Component for CardDoc {
-        type Message = ();
-        type Properties = ();
-
-        fn create(_: Self::Properties, _link: ComponentLink<Self>) -> Self {
-            CardDoc {}
-        }
-
-        fn update(&mut self, _msg: Self::Message) -> ShouldRender {
-            true
-        }
-
-        fn change(&mut self, _props: Self::Properties) -> ShouldRender {
-            true
-        }
-
-        fn view(&self) -> Html {
-            let source = crate::include_example!("example.rs");
-
-            html! {
-                <div>
-                    <h1>{"Card"}</h1>
-                    <div>{source}</div>
-                </div>
-            }
         }
     }
 }
